refactor(Volt): Reorganize drivetrains
