+++ Volt/src/main/kotlin/dev/kingssack/volt/integrations/rr/localizer/RoadRunnerLocalizer.kt
@@ -1,10 +1,9 @@
-package dev.kingssack.volt.util
+package dev.kingssack.volt.integrations.rr.localizer
 
 import com.acmerobotics.roadrunner.Pose2d
 import com.acmerobotics.roadrunner.PoseVelocity2d
 
-
-interface Localizer {
+interface RoadRunnerLocalizer {
     var pose: Pose2d
 
     fun update(): PoseVelocity2d
